jetstream client helper
This removes the old JetStreamMgmt object and makes a more generic helper
thats a bit better structured.

jsm utility updated to use it and old one removed.

This new set of code uses Stream and Consumer names, I'll merge when JS is
done renaming. Though earlier is fine too as the jsm utility kept old names
all through so it should be invisible to users.