Add a --tls-dump-certs flag
Add a diagnostic flag `--tls-dump-certs` which sets up a callback so that on
every TLS connection establishment, we will dump out voluminous information
about the certificates, including:

 * the chains constructed by the client from the certs presented by the server
 * the PEM data for each certificate presented by the server, preceded by a
   label giving the name and expiration data, and any subjectAltName DNSNames

This should be sufficient to eyeball the most pertinent data while also
providing the full certs for further analysis if needed.

The diagnostics about counts etc are emitted using the log package, per
existing practice, but the certificates I put on stdout to ease piping to
analysis tools.