Humanize byte arguments on the cli and prompts
This uses a bytes parser derived from, and compatible with,
the server options parser but additionally supports TB, tb and
t as aliases for T. All are 1024 based.

We use that in cli arguments and update all other bytes
parsing to be the same, this means 1GB is consistent with
the rendering of 1GiB - all 1024 based.

Do the same on stream, kv and object.

Fixes a bug that required a size be set on accounts
that do not require it and default to 256MB when it is
required

Signed-off-by: R.I.Pienaar <rip@devco.net>