Improve help output
This uses master of kingpin to get some additional
features around templates, use those to inject some
helper function into the templates and create a new
help template

The basic idea is that just typing nats must give a
overview and doing --help on any sub command will show
flags (including common ones) and more

This is a quick stop gap, we can do better but for now
this is already a huge step up.  Technically a breaking
change due to making it incompatible with the blessed
kingpin, but meh.

Signed-off-by: R.I.Pienaar <rip@devco.net>