handle memory streams better in account backup
These will now be treated as warnings since there is nothing that a user
can do about them - we do not support skipping certain streams etc.

So they will now log, and then report in a seperate section for warnings
and not cause a 1 exit code.

However they can be treated as critical using --critical-warnings

Signed-off-by: R.I.Pienaar <rip@devco.net>