Rework nats server run
The behavior of always cleaning contexts after each
run and creating new random passwords and ports on
each start was too difficult to use, especially given
the lack of natscontext support in any other languages.

Now when the required contexts exist the passwords and
port is read from them and reused.  Cleaning behavior
can be enabled with --clean

Signed-off-by: R.I.Pienaar <rip@devco.net>