New sub-command: account tls
Running `nats account tls` will report all the valid TLS chains verified
by the client for the current TLS connection.  Each chain shows the PEM-encoded
form of each cert in the chain, preceded by some hopefully-useful data as
comments.  Add `--no-pem` to skip the PEM and leave the comments.

We'll warn, and exit with an error, if we're within 1 week of the expiration of
a certificate.  This is adjustable with a flag.  Exiting if a cert has expired
is not adjustable (but shouldn't happen with verified chains).

There's logic for OCSP verification, taken from my smtpdane tool (I'm the
copyright holder of that tool so this is fine, relicensing this copied 20 line
snippet as part of natscli).