Removing the defer to delete the durable consumer in bench js consume and fetch, and rely only on the idle timeout to cleanup the consumer instead.
This is because otherwise the consumer delete happens while there are still some acks already published by the client(s) but not yet processed by JS. When using an interest or working queue stream this means that (even with a relatively small batch of 100) some messages do not get removed from the stream after the bench consume or fetch run is finished.

Note: having to rely on the idle timeout for consumer cleanup is still not 'perfect' (in some cases the idle timeout value may not be large enough), but this seems to be the only option available at the moment.
Signed-off-by: Jean-Noël Moyne <jnmoyne@gmail.com>