Bench command in multisubject mode subject format improvement
When using nats bench in multisubject mode to generate messages in order to evaluate the amount of memory used by the indexes according to the number of subjects being stored in the stream it is more accurate if all the subjects have the same size. This achieves that by using padding 0 up to the length of the max number of subjects when creating the subjects.

Signed-off-by: Jean-Noël Moyne <jnmoyne@gmail.com>