(#12633) Fix edit when not passing --compression flag
In the past when a user ran `stream edit` without the `--compression` flag, the
empty flag would get parsed as the compression type `none`. This means that if
compression was set to `s2`, every edit without the flag would look like there
was a change to the stream's compression.

We now use `isSetByUser` to determine if the value has changed before trying
to parse a value from the cli.