nats server check credential --context-credential
A common use-case for checking credentials is "check my current credentials".
Rather than force users to manually grab the path from their context, provide a
flag `--context-credential` to just use the credential from that.

Changes in existing behavior: a different error flow for missing `--credential`
when this new flag is not given.

I couldn't see a way in the fisk library to have a mandatory
mutually-exclusive flag group, so coded up the checks manually when dropping
`Required()` from `--credential`.